chore(site): apps cannot offer reduced prices
in discord's developer terms effective october 7 2024, apps that offer premium features are required to use the premium apps program and must not charge a higher price on discord than on other platform (e.g. ko-fi).

in response we've raised the price on other platforms for now to match what we upcharged on discord, initially only to combat discord's 33% cut from every purchase.