feat(bot): backup components on ephemeral messages
components now operate on an ID-only basis. if a component is used on an ephemeral message (specifically one that is a response to another component), a new durable object will be created for it which lasts one hour. this change also allows users to "manipulate" messages owned by the bot, but access is guild-restricted so there should be no damage.